Titan on 32GB SD Card
After Titan was made crowetic proof, shit has calmed down and his goofy stuff is not impacting it anymore. 
Disk I/O is due to the fact that fucker still runs on a slow SD card. Jack is too lazy to put it for me on SSD for more stress testing.